added postgres for register and login

+++ src/main/java/edu/iu/romanju/primesservice/service/IPrimesService.java
@@ -1,5 +1,5 @@
 package edu.iu.romanju.primesservice.service;
 
 public interface IPrimesService {
-    boolean isPrime(int n);
+    boolean isPrime(long n);
 }
