+++ src/main/java/edu/iu/romanju/primesservice/service/PrimesService.java
@@ -3,17 +3,19 @@ package edu.iu.romanju.primesservice.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class PrimesService implements IPrimesService{
+public class PrimesService implements IPrimesService {
     @Override
-    public boolean isPrime(int n){
-        if (n == 2){
+    public boolean isPrime(long n) {
+        if (n == 2) {
             return true;
         }
-        for(long i = 2L; i<n ; i++){
-            if(n%i==0) {
+
+        for (long i = 2L; i < n; i++) {
+            if (n % i == 0) {
                 return false;
             }
         }
+
         return true;
     }
 }
